Move common components to a separate package
