+++ app/src/main/java/com/flaringapp/compose/topbar/ui/samples/common/SampleTopBarImage.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.flaringapp.compose.topbar.ui.samples
+package com.flaringapp.compose.topbar.ui.samples.common
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.height
